% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ihs}
\alias{ihs}
\title{Function for taking the inverse hyperbolic sine of a number}
\usage{
ihs(x)
}
\arguments{
\item{x}{a double}
}
\value{
a double
}
\description{
Function for taking the inverse hyperbolic sine of a number
}
