% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{robustses}
\alias{robustses}
\title{Calculates robust standard errors from a regression model}
\usage{
robustses(x, c, setype = "HC1")
}
\arguments{
\item{x}{a model from lm or similar}

\item{c}{the clustering variable}

\item{setype}{standard error types}
}
\value{
the variance
}
\description{
Calculates robust standard errors from a regression model
}
