% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sumstats.wtd}
\alias{sumstats.wtd}
\title{Creates a summary statistics from a data frame with weights}
\usage{
sumstats.wtd(df, wts, labs = NULL, d = 2)
}
\arguments{
\item{df}{a data frame}

\item{wts}{vector of weights same length as the data frame}

\item{labs}{vector of variable names - defaults to column names of df}
}
\value{
the variance
}
\description{
Creates a summary statistics from a data frame with weights
}
