% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weighted.var}
\alias{weighted.var}
\title{Calculates the weighted variance of a vector}
\usage{
weighted.var(x, w)
}
\arguments{
\item{x}{a vector}

\item{w}{a vector of weights same length as x}
}
\value{
the variance
}
\description{
Calculates the weighted variance of a vector
}
